requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=h6b3803e_1
absl-py=1.2.0=pyhd8ed1ab_0
aiohttp=3.8.1=py310hf8d0d8f_1
aiosignal=1.2.0=pyhd8ed1ab_0
alembic=1.8.1=pypi_0
aom=3.5.0=h7ea286d_0
appnope=0.1.3=pyhd8ed1ab_0
astor=0.8.1=pypi_0
asttokens=2.0.5=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
autograd=1.4=pypi_0
autograd-gamma=0.5.0=pypi_0
autopage=0.5.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blinker=1.4=py_1
brotli=1.0.9=pypi_0
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py310hf8d0d8f_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2023.7.22=hf0a4a13_0
cachetools=4.2.4=pypi_0
captum=0.5.0=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310he00a5c5_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.0.4=pypi_0
cliff=4.0.0=pypi_0
cloudpickle=2.2.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.4.2=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
colorlog=6.7.0=pypi_0
contourpy=1.0.5=py310h2887b22_0
cryptography=37.0.1=py310h834c97f_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.0=py310h1105856_0
decorator=5.1.1=pyhd8ed1ab_0
distlib=0.3.6=pypi_0
ecos=2.0.11=py310hf1a086a_0
eli5=0.13.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
et-xmlfile=1.1.0=pypi_0
executing=0.9.1=pyhd8ed1ab_0
expat=2.4.9=hb7217d7_0
fastjsonschema=2.16.2=pypi_0
feather-format=0.4.1=pypi_0
ffmpeg=5.1.2=gpl_hcccd11d_101
filelock=3.8.0=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h82840c6_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.34.4=pypi_0
formulaic=0.4.0=pypi_0
freetype=2.12.1=hd633e50_0
frozenlist=1.3.1=py310h8c01e39_0
fsspec=2022.7.1=pypi_0
future=0.18.2=py310hbe9552e_5
gettext=0.19.8.1=h0186832_1009
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.8=h9f1a10d_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.1=py_2
grpc-cpp=1.46.4=hbdbd451_3
grpcio=1.46.4=py310ha83b316_3
h5py=3.7.0=pypi_0
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310hbe9552e_0
inflate64=0.3.0=pypi_0
interface-meta=1.3.0=pypi_0
ipykernel=6.15.1=pyh736e0ef_0
ipython=8.4.0=py310hbe9552e_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=he4db4b2_2
jsonschema=4.16.0=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=4.11.1=py310hbe9552e_0
kiwisolver=1.4.4=py310hd23d0e8_0
lame=3.100=h1a8c8d9_1003
lcms2=2.12=had6a04f_0
lerc=4.0.0=h9a09cb3_0
libabseil=20211102.0=cxx17_hc0b94f3_2
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=16_osxarm64_openblas
libcxx=14.0.6=h04bba0f_0
libdeflate=1.13=h57fd34a_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.17=he4db4b2_0
libidn2=2.3.3=he4db4b2_0
liblapack=3.9.0=16_osxarm64_openblas
libopenblas=0.3.21=openmp_hcb59c3b_0
libosqp=0.6.3=h313beb8_0
libpng=1.6.38=h76d750c_0
libprotobuf=3.20.1=h332123e_0
libqdldl=0.1.7=hb7217d7_0
libsodium=1.0.18=h27ca646_1
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.4.0=hd6ca8d7_3
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.14=h9d8dfc2_4
libzlib=1.2.12=ha287fd2_2
lifelines=0.27.1=pypi_0
llvm-openmp=16.0.6=h1c12783_0
llvmlite=0.39.0=pypi_0
mako=1.2.3=pypi_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310hf8d0d8f_1
matplotlib=3.5.3=pypi_0
matplotlib-base=3.6.0=py310h78c5c2f_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
msgpack=1.0.4=pypi_0
multidict=6.0.2=py310hf8d0d8f_1
multivolumefile=0.2.3=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nbformat=5.7.3=pypi_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
ninja=1.11.0=hf86a087_0
numba=0.56.0=pypi_0
numexpr=2.8.4=py310h2b830bf_0
numpy=1.21.0=pypi_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.3.1=hb7217d7_1
openjpeg=2.5.0=h5d4e404_1
openpyxl=3.1.0=pypi_0
openssl=3.0.5=h03a7124_2
optuna=3.0.3=pypi_0
osqp=0.6.3=py310h46d7db6_0
p11-kit=0.24.1=h29577a5_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pbr=5.10.0=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310h173adc3_2
pip=22.2.2=pyhd8ed1ab_0
platformdirs=2.5.2=pypi_0
plotly=5.10.0=pypi_0
prettytable=3.4.1=pypi_0
prompt-toolkit=3.0.30=pyha770c72_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=py310h02f21da_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py7zr=0.20.0=pypi_0
pyarrow=9.0.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pybcj=1.0.1=pypi_0
pycox=0.2.3=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodomex=3.15.0=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.12.0=pyhd8ed1ab_0
pyjwt=2.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyperclip=1.8.2=pypi_0
pyppmd=0.18.3=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py310hbe9552e_5
python=3.10.5=h4eee789_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-graphviz=0.20.1=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_0
pytorch-lightning=1.7.1=pypi_0
pytorch-lightning-bolts=0.3.2.post1=pypi_0
pytz=2022.2.1=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py310hf8d0d8f_4
pyzmq=23.2.0=py310hede2015_0
pyzstd=0.15.3=pypi_0
qdldl-python=0.1.7=py310h46d7db6_0
ray=2.0.0=pypi_0
re2=2022.06.01=h9a09cb3_0
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.2=pypi_0
scikit-survival=0.21.0=py310ha66af24_0
scipy=1.8.1=py310hdb41229_2
seaborn=0.12.2=pypi_0
setuptools=64.0.0=py310hbe9552e_0
shap=0.41.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn-pandas=2.2.0=pypi_0
sleef=3.5.1=h156473d_2
slicer=0.0.7=pypi_0
sqlalchemy=1.4.41=pypi_0
sqlite=3.39.2=h40dfcc0_0
stack_data=0.3.0=pyhd8ed1ab_0
stevedore=4.0.1=pypi_0
svt-av1=1.2.1=he23bcac_0
tabulate=0.8.10=pypi_0
tenacity=8.1.0=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorboardx=2.5.1=pypi_0
texttable=1.6.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=he1e0b03_0
torchaudio=0.12.1=py310_cpu
torchmetrics=0.9.3=pypi_0
torchtuples=0.2.2=pypi_0
torchvision=0.13.1=py310_cpu
tornado=6.2=py310h02f21da_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.3.0=pyhd8ed1ab_0
typing-extensions=4.3.0=hd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022b=h191b570_0
unicodedata2=14.0.0=py310hf8d0d8f_1
urllib3=1.26.11=pyhd8ed1ab_0
virtualenv=20.16.5=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
yaml=0.2.5=h3422bc3_2
yarl=1.7.2=py310hf8d0d8f_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=ha287fd2_2
zstd=1.5.2=h8128057_4
